+++ src/on_computer/pathfinding/PathfindingAlgorithm.py
